tRPC: User asset breakdown (#2728)
* include verified assets in user assets sort

* more descriptive naming

* fix default sort of balances

* comment

* query invalidation

* clean up default sort

* user total assets procedure

* use breakdown instead

* invalidate

* more readable balances

* fix assets table scroll

* @JoseRFelix review

* include coins from locks as pooled

* naming and code org

* naming and fix get delegations
