+++ packages/web/server/queries/complex/assets/market.ts
@@ -6,10 +6,14 @@ import { LRUCache } from "lru-cache";
 import { DEFAULT_LRU_OPTIONS } from "~/config/cache";
 import { AssetLists } from "~/config/generated/asset-lists";
 
-import { queryTokenMarketCaps } from "../../imperator";
+import { queryCoingeckoCoin } from "../../coingecko";
+import {
+  queryAllTokenData,
+  queryTokenMarketCaps,
+  TokenData,
+} from "../../imperator";
 import { Asset, AssetFilter, getAssets } from ".";
 import { DEFAULT_VS_CURRENCY } from "./config";
-import { getAssetData, getAssetMarketCapRank } from "./info";
 import { getAssetPrice } from "./price";
 
 export type AssetMarketInfo = Partial<{
@@ -21,7 +25,7 @@ export type AssetMarketInfo = Partial<{
 
 const marketInfoCache = new LRUCache<string, CacheEntry>(DEFAULT_LRU_OPTIONS);
 /** Cached function that returns an asset with market info included. */
-export async function getAssetMarketInfo<TAsset extends Asset>({
+export async function getMarketAsset<TAsset extends Asset>({
   asset,
 }: {
   asset: TAsset;
@@ -41,7 +45,8 @@ export async function getAssetMarketInfo<TAsset extends Asset>({
           return undefined;
         })
       )?.find((mCap) => mCap.symbol === asset.coinDenom)?.market_cap;
-      const priceChange24h = (await getAssetData(asset))?.price_24h_change;
+      const priceChange24h = (await getAssetMarketActivity(asset))
+        ?.price_24h_change;
       const marketCapRank = await getAssetMarketCapRank(asset);
 
       return {
@@ -64,7 +69,7 @@ export async function getAssetMarketInfo<TAsset extends Asset>({
 
 /** Maps and adds general supplementary market data such as current price and market cap to the given type.
  *  If no assets provided, they will be fetched and passed the given search params. */
-export async function mapGetAssetMarketInfos<TAsset extends Asset>({
+export async function mapGetMarketAssets<TAsset extends Asset>({
   assetList = AssetLists,
   ...params
 }: {
@@ -74,7 +79,59 @@ export async function mapGetAssetMarketInfos<TAsset extends Asset>({
   let { assets } = params;
   if (!assets) assets = (await getAssets(params)) as TAsset[];
 
-  return await Promise.all(
-    assets.map((asset) => getAssetMarketInfo({ asset }))
-  );
+  return await Promise.all(assets.map((asset) => getMarketAsset({ asset })));
+}
+
+const assetMarketCache = new LRUCache<string, CacheEntry>(DEFAULT_LRU_OPTIONS);
+
+/** Gets the numerical market cap rank given a token symbol/denom.
+ *  Returns `undefined` if a market cap is not available for the given symbol/denom. */
+async function getAssetMarketCapRank({
+  coinGeckoId,
+}: {
+  coinGeckoId: string | undefined;
+}): Promise<number | undefined> {
+  if (!coinGeckoId) return;
+
+  return await cachified({
+    cache: assetMarketCache,
+    ttl: 1000 * 60 * 15, // 15 minutes since market ranks don't change often
+    key: "market-cap-" + coinGeckoId,
+    getFreshValue: async () => {
+      try {
+        const coinGeckoCoin = await queryCoingeckoCoin(coinGeckoId);
+
+        return coinGeckoCoin.market_cap_rank;
+      } catch {
+        // ignore error and return undefined, since market cap rank is non-critical
+      }
+    },
+  });
+}
+
+/** Fetches general asset info such as price and price change, liquidity, volume, and name
+ *  configured outside of our asset list (from data services).
+ *  Returns `undefined` for a given coin denom if there was an error or it's not available. */
+async function getAssetMarketActivity({ coinDenom }: { coinDenom: string }) {
+  const assetMarketMap = await cachified({
+    cache: assetMarketCache,
+    ttl: 1000 * 60 * 5, // 5 minutes since there's price data
+    key: "allTokenData",
+    getFreshValue: async () => {
+      try {
+        const allTokenData = await queryAllTokenData();
+
+        const tokenInfoMap = new Map<string, TokenData>();
+        allTokenData.forEach((tokenData) => {
+          tokenInfoMap.set(tokenData.symbol, tokenData);
+        });
+        return tokenInfoMap;
+      } catch (error) {
+        console.error("Could not fetch token infos", error);
+        return new Map<string, TokenData>();
+      }
+    },
+  });
+
+  return assetMarketMap.get(coinDenom);
 }
